+++ src/main/java/org/sonar/plugins/delphi/antlr/ast/node/ArgumentListNode.java
@@ -36,6 +36,7 @@ public final class ArgumentListNode extends DelphiNode {
       for (int i = 0; i < jjtGetNumChildren(); ++i) {
         imageBuilder.append(jjtGetChild(i).getImage());
       }
+      imageBuilder.append(")");
       image = imageBuilder.toString();
     }
     return image;