+++ src/test/java/org/sonar/plugins/delphi/pmd/rules/ExtraneousArgumentListCommasRuleTest.java
@@ -1,0 +1,39 @@
+package org.sonar.plugins.delphi.pmd.rules;
+
+import static org.sonar.plugins.delphi.utils.conditions.RuleKeyAtLine.ruleKeyAtLine;
+
+import org.junit.Test;
+import org.sonar.plugins.delphi.utils.builders.DelphiTestUnitBuilder;
+
+public class ExtraneousArgumentListCommasRuleTest extends BasePmdRuleTest {
+
+  @Test
+  public void testArgumentListWithoutTrailingCommaShouldNotAddIssue() {
+    DelphiTestUnitBuilder builder =
+        new DelphiTestUnitBuilder()
+            .appendImpl("procedure Test;")
+            .appendImpl("begin")
+            .appendImpl("  Foo(1, 2, 3);")
+            .appendImpl("end;");
+
+    execute(builder);
+
+    assertIssues().isEmpty();
+  }
+
+  @Test
+  public void testArgumentListWithTrailingCommaShouldAddIssue() {
+    DelphiTestUnitBuilder builder =
+        new DelphiTestUnitBuilder()
+            .appendImpl("procedure Test;")
+            .appendImpl("begin")
+            .appendImpl("  Foo(1, 2, 3,);")
+            .appendImpl("end;");
+
+    execute(builder);
+
+    assertIssues()
+        .hasSize(1)
+        .areExactly(1, ruleKeyAtLine("ExtraneousArgumentListCommasRule", builder.getOffset() + 3));
+  }
+}